Issue-3504: Fixed code review notes
